clean up for PR
