+++ frontend/App.js
@@ -1,20 +1,40 @@
-import 'react-native-gesture-handler';
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
-import { StyleSheet, Text, View, SafeAreaView, TouchableOpacity, Image, Button, Alert } from 'react-native';
+import { StyleSheet, Text, View, TouchableOpacity } from 'react-native';
 import Constants from 'expo-constants';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
 
-import MainStackNavigator from './pod_components/MainStackNavigator.js';
+import HomeScreen from './pod_components/HomeScreen.js';
+import TraineePodScreen from './pod_components/TraineePod.js';
+import AssociatePodScreen from './pod_components/AssociatePod.js';
+import PartnerPodScreen from './pod_components/PartnerPod.js';
 
-export default function App() {  
-    const [data, setData] = React.useState("default");
-    fetch(Constants.manifest.extra.apiUrl).then(rawResult => rawResult.json()).then(jsonResult => {
-        setData(jsonResult.data) 
-    });
-        
+const Stack = createStackNavigator();
+
+export default function MainStackNavigator() {          
     return (
-       <MainStackNavigator />
+       <NavigationContainer>
+           <Stack.Navigator initialRouteName="Home">
+               <Stack.Screen name="Home" component={HomeScreen} />
+               <Stack.Screen name="Trainee Pod" component={TraineePodScreen} />
+               <Stack.Screen name="Associate Pod" component={AssociatePodScreen} />
+               <Stack.Screen name="Partner Pod" component={PartnerPodScreen} />
+               <Stack.Screen name="Random Screen" component={RandomScreen} />
+            </Stack.Navigator>
+        </NavigationContainer>
     );
+}
+
+function RandomScreen() {
+  return (
+    <View style={styles.container}>
+      <Text style={styles.title}>
+          When the competencies screen, career pathway screen, and life 
+          essential screens get set up, replace this screen with them.
+      </Text>
+    </View>
+  );
 }
 
 const styles = StyleSheet.create({
@@ -23,5 +43,10 @@ const styles = StyleSheet.create({
         backgroundColor: '#fff',
         alignItems: 'center',
         justifyContent: 'center',
-    }
+    },
+    title: {
+        fontSize: 30,
+        marginTop: 20,
+        fontWeight: 'bold',
+    },
 });
